+++ test/summary.spec.js
@@ -2,6 +2,10 @@ import Summary from "../src/summary"
 
 let log;
 
+const opts = {
+  name: "My Server"
+}
+
 describe("Summary", () => {
   beforeEach(() => {
     log = td.replace(console, "log")
@@ -12,8 +16,15 @@ describe("Summary", () => {
   })
 
   describe("#print", () => {
+    it("print the server name", () => {
+      const summary = new Summary([], opts)
+      summary.print()
+
+      td.verify(log(td.matchers.contains(opts.name)))
+    })
+
     it("prints nothing when there are no new tapes and no unused tapes", () => {
-      const summary = new Summary([])
+      const summary = new Summary([], opts)
 
       summary.print()
 
@@ -26,7 +37,7 @@ describe("Summary", () => {
         {new: true, used: true, path: "path1"},
         {used: true, path: "path2"},
         {new: true, used: true, path: "path3"}
-      ])
+      ], opts)
 
       summary.print()
 
@@ -40,7 +51,7 @@ describe("Summary", () => {
         {path: "path1"},
         {used: true, path: "path2"},
         {path: "path3"}
-      ])
+      ], opts)
 
       summary.print()